Make `connection_intent` client-specific
