+++ rust/connlib/tunnel/src/lib.rs
@@ -45,7 +45,7 @@ pub use control_protocol::Request;
 pub use role_state::{ClientState, GatewayState};
 pub use webrtc::peer_connection::sdp::session_description::RTCSessionDescription;
 
-use crate::ip_packet::MutableIpPacket;
+use crate::ip_packet::{IpPacket, MutableIpPacket};
 use crate::role_state::RoleState;
 use connlib_shared::messages::SecretKey;
 use index::IndexLfsr;
@@ -628,6 +628,71 @@ where
                 }
             }));
     }
+
+    #[inline(always)]
+    fn connection_intent(self: &Arc<Self>, packet: IpPacket<'_>) {
+        const MAX_SIGNAL_CONNECTION_DELAY: Duration = Duration::from_secs(2);
+
+        // We can buffer requests here but will drop them for now and let the upper layer reliability protocol handle this
+        if let Some(resource) = self.get_resource(packet.source()) {
+            // We have awaiting connection to prevent a race condition where
+            // create_peer_connection hasn't added the thing to peer_connections
+            // and we are finding another packet to the same address (otherwise we would just use peer_connections here)
+            let mut awaiting_connection = self.awaiting_connection.lock();
+            let conn_id = ConnId::from(resource.id());
+            if awaiting_connection.get(&conn_id).is_none() {
+                tracing::trace!(
+                    resource_ip = %packet.destination(),
+                    "resource_connection_intent",
+                );
+
+                awaiting_connection.insert(conn_id, Default::default());
+                let dev = Arc::clone(self);
+
+                let mut connected_gateway_ids: Vec<_> = dev
+                    .gateway_awaiting_connection
+                    .lock()
+                    .clone()
+                    .into_keys()
+                    .collect();
+                connected_gateway_ids
+                    .extend(dev.resources_gateways.lock().values().collect::<Vec<_>>());
+                tracing::trace!(
+                    gateways = ?connected_gateway_ids,
+                    "connected_gateways"
+                );
+                tokio::spawn(async move {
+                    let mut interval = tokio::time::interval(MAX_SIGNAL_CONNECTION_DELAY);
+                    loop {
+                        interval.tick().await;
+                        let reference = {
+                            let mut awaiting_connections = dev.awaiting_connection.lock();
+                            let Some(awaiting_connection) =
+                                awaiting_connections.get_mut(&ConnId::from(resource.id()))
+                            else {
+                                break;
+                            };
+                            if awaiting_connection.response_received {
+                                break;
+                            }
+                            awaiting_connection.total_attemps += 1;
+                            awaiting_connection.total_attemps
+                        };
+                        if let Err(e) = dev
+                            .control_signaler
+                            .signal_connection_to(&resource, &connected_gateway_ids, reference)
+                            .await
+                        {
+                            // Not a deadlock because this is a different task
+                            dev.awaiting_connection.lock().remove(&conn_id);
+                            tracing::error!(error = ?e, "start_resource_connection");
+                            let _ = dev.callbacks.on_error(&e);
+                        }
+                    }
+                });
+            }
+        }
+    }
 }
 
 impl<C, CB> Tunnel<C, CB, GatewayState>
@@ -657,7 +722,8 @@ where
                 let mut buf = [0u8; MAX_UDP_SIZE];
                 loop {
                     let Some(packet) = device.read().await? else {
-                        return Ok(());
+                        // Reading a bad IP packet or otherwise from the device seems bad. Should we restart the tunnel or something?
+                        return Result::Ok(());
                     };
 
                     let dest = packet.destination();
